Move identifier into a separate mod
